ipl/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
